RDK-53282: Merge rdk_43243_cdai_ctsb branch to dev_sprint_24_2 branch
Reason for change: Squashed commit for sprint merge from rdk_43243_cdai_ctsb
Modified duration calculation to start from presentationTimeOffset instead of t value of timeline index start.
Adjusted the logic in fragmentcollector_mpd.cpp to ensure accurate duration computation.
Address comments to LLD CDAI with CTSB implementation
Fix for regression in multiple ads added in
a single adbreak with L1. Fix for regression with L2 test 8004
Fix for Seamless audio switch in FetcherLoopNew
Test Procedure: Check L1 and L2 pass successfully and CDAI playback
Risks: Low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_24_2
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_24_2
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_24_2

(cherry picked from commit 7ac0110af22d8d302dab0d89e59cb4fb4493df8f)

+++ isobmff/isobmffbox.cpp
@@ -494,7 +494,7 @@ void TfdtBox::setBaseMDT(uint64_t mdt)
 		}
 		else
 		{
-			WRITE_U32(baseMDT_loc, mdt);
+			WRITE_U32(baseMDT_loc, static_cast<uint32_t>(mdt));
 		}
 	}
 }
